Support for JSHint reporters.
Similar usage as formatters in grunt-contrib-csslint.

It emulates the official reporter design so existing reporters work. I
overwrite `process.stdout` to pipe the reports to file.

In the process I fixed an issue where the force option was overwritten
when jshintrc is specified, causing the task to succeed when using
jshintrc.